requirements.txt
keras==2.13.1
openai==0.28.0
Flask==2.3.3
numpy==1.24.1
Pillow==9.3.0
tensorflow==2.13.0
Flask-Cors==4.0.0